Redesign. Add statusbar. Fix bugs. Add new functional. Add stylesheet resource

+++ src/FocuseSpace/notes.cpp
@@ -30,52 +30,47 @@ Notes::Notes(QMainWindow *parent) :
     else {
         qDebug() << "Таблица notes создана";
     }
+    setFixedSize(800, 650);
 
-    setWindowTitle("~ notes ~");
-    setFixedSize(800, 720);
-
-    title = new QLabel("Notes", this);
-    title->setGeometry(0, 20, 800, 50);
-    title->setAlignment(Qt::AlignCenter);
-    title->setFont(QFont("SF Pro Black", 30));
-    title->setStyleSheet( " color: #ffffff; " );
+    statusBar = new QStatusBar(this);
 
     noteName = new QLineEdit(this);
-    noteName->setGeometry(280, 80, 500, 50);
+    noteName->setGeometry(280, 10, 500, 50);
     noteName->setFont(QFont("SF Pro Black", 14));
     noteName->setStyleSheet( " background-color: #393939; selection-background-color: #999999; selection-color: #ffffff; color: #ffffff; border-width: 2px; border-style: solid; border-radius: 10px; border-color: #393939; alternate-background-color: #303030;" );
     noteName->setPlaceholderText("Title");
 
     noteEdit = new QTextEdit(this);
-    noteEdit->setGeometry(280, 150, 500, 500);
+    noteEdit->setGeometry(280, 80, 500, 500);
     noteEdit->setStyleSheet( " background-color: #393939; selection-background-color: #999999; selection-color: #ffffff; color: #ffffff; border-width: 2px; border-style: solid; border-radius: 10px; border-color: #393939; alternate-background-color: #303030;" );
     noteEdit->setFont(QFont("SF Pro Black", 11));
     noteEdit->setPlaceholderText("Note");
 
     notesList = new QListWidget(this);
-    notesList->setGeometry(20, 80, 240, 570);
+    notesList->setGeometry(20, 10, 240, 570);
     notesList->setStyleSheet( " background-color: #393939; selection-background-color: #999999; selection-color: #ffffff; color: #ffffff; border-width: 2px; border-style: solid; border-radius: 10px; border-color: #393939; alternate-background-color: #303030;" );
     notesList->setFont(QFont("SF Pro Black", 12));
     connect(notesList, SIGNAL(itemDoubleClicked(QListWidgetItem*)), this, SLOT(doubleClick(QListWidgetItem*)));
 
     saveButton = new QPushButton("SAVE", this);
-    saveButton->setGeometry(480, 665, 295, 40);
+    saveButton->setGeometry(480, 590, 300, 40);
     saveButton->setStyleSheet( " background-color: #444444; color: #ffffff; border-width: 2px; border-style: solid; border-radius: 10px; border-color: #444444; alternate-background-color: #303030;" );
     saveButton->setFont(QFont("SF Pro Black", 10));
     connect(saveButton, SIGNAL(clicked()), this, SLOT(saveNotes()));
 
     backButton = new QPushButton("BACK", this);
-    backButton->setGeometry(20, 665, 235, 40);
+    backButton->setGeometry(20, 590, 240, 40);
     backButton->setStyleSheet( " background-color: #444444; color: #ffffff; border-width: 2px; border-style: solid; border-radius: 10px; border-color: #444444; alternate-background-color: #303030;" );
     backButton->setFont(QFont("SF Pro Black", 10));
     connect(backButton, SIGNAL(clicked()), this, SLOT(toMainWindow()));
 
     removeButton = new QPushButton("REMOVE", this);
-    removeButton->setGeometry(280, 665, 175, 40);
+    removeButton->setGeometry(280, 590, 185, 40);
     removeButton->setStyleSheet( " background-color: #444444; color: #ffffff; border-width: 2px; border-style: solid; border-radius: 10px; border-color: #444444; alternate-background-color: #303030;" );
     removeButton->setFont(QFont("SF Pro Black", 10));
     connect(removeButton, SIGNAL(clicked()), this, SLOT(removeNote()));
 
+    setStatusBar(statusBar);
 
     query.exec("SELECT * FROM notes");
     while (query.next()) {
@@ -89,6 +84,7 @@ Notes::Notes(QMainWindow *parent) :
         notesList->addItem(item);
 
         qDebug() << name << " was loaded";
+        statusBar->showMessage("All notes was loaded");
     }
 }
 
@@ -120,6 +116,7 @@ void Notes::saveNotes() {
         }
         else {
             qDebug() << title << " was saved";
+            statusBar->showMessage(title + " was saved");
         }
 
         noteEdit->clear();
@@ -130,33 +127,39 @@ void Notes::saveNotes() {
     }
     else {
         qDebug() << "Note is empty";
+        statusBar->showMessage("Note is empty");
     }
 }
 
 void Notes::doubleClick(QListWidgetItem *item) {
-    QSqlQuery query;
+    QString text = noteEdit->toPlainText();
+    if(text.isEmpty()) {
+        QSqlQuery query;
 
-    QListWidgetItem *selectedItem = notesList->currentItem();
-    
-    QString note = selectedItem->text();
+        QListWidgetItem *selectedItem = notesList->currentItem();
+        
+        QString note = selectedItem->text();
 
-    QString name = selectedItem->text();
+        QString name = selectedItem->text();
 
-    query.exec("SELECT * FROM notes WHERE name = '"+name+"'");
+        query.exec("SELECT * FROM notes WHERE name = '"+name+"'");
 
-    if (query.exec()) {
-        while (query.next()) {
-            QString name = query.value("name").toString();
-            QString text = query.value("text").toString();
+        if (query.exec()) {
+            while (query.next()) {
+                QString name = query.value("name").toString();
+                QString text = query.value("text").toString();
 
-            noteName->setText(name);
-            noteEdit->setText(text);
+                noteName->setText(name);
+                noteEdit->setText(text);
 
-            qDebug() << name << " was opened";
+                qDebug() << name << " was opened";
+                statusBar->showMessage(name + " was opened");
+            }
         }
-    }else {
-        qDebug() << "Error moving data: " << query.lastError().text();
-    }
+        else {
+            qDebug() << "Error moving data: " << query.lastError().text();
+            statusBar->showMessage(query.lastError().text());
+        }
 
 
         query.prepare("DELETE FROM notes WHERE name = '"+name+"'");
@@ -164,9 +167,16 @@ void Notes::doubleClick(QListWidgetItem *item) {
         if (query.exec()) {
             delete notesList->takeItem(notesList->row(item));
             qDebug() << name << " was opened";
+            statusBar->showMessage(name + " was opened");
         } else {
             qDebug() << "Error moving data: " << query.lastError().text();
+            statusBar->showMessage(query.lastError().text());
         }
+    }
+    else {
+        qDebug() << "Previous changes were not saved. save them and try again.";
+        statusBar->showMessage("Previous changes were not saved. save them and try again.");
+    }
 }
 
 
@@ -182,7 +192,9 @@ void Notes::removeNote() {
     if (query.exec()) {
         delete selectedItem;
         qDebug() << name << " was deleted";
+        statusBar->showMessage(name + " was deleted");
     } else {
         qDebug() << "Error deleting data:" << query.lastError().text();
+        statusBar->showMessage(query.lastError().text());
     }
 }
